Orientations and TRS completed.
Model Loader loads the Translations, Rotations, Scales and TRS matrices from the glTF.
The defaults are corrected to actual default.
NOTE: glm::quat(w,x,y,z) ; glm::make_quat({x,y,z,w});
The OrientationTest and BoomBoxWithAxis both work.

+++ Blaze/Renderer.cpp
@@ -421,7 +421,7 @@ namespace blaze
 
 		VkViewport viewport = {};
 		viewport.x = 0.0f;
-		viewport.y = swapchainExtent.get().height;
+		viewport.y = static_cast<float>(swapchainExtent.get().height);
 		viewport.width = (float)swapchainExtent.get().width;
 		viewport.height = -(float)swapchainExtent.get().height;
 		viewport.minDepth = 0.0f;
